Crate tlspuffin

TODO: Write intro: https://gitlab.inria.fr/mammann/tlspuffin/-/issues/64

Used protocol and cryptographic libraries

In order to easily implement concrete functions, we use several libraries which provide us with predefined encoders for TLS packets, cryptographic primitives, as well as higher level cryptographic operations specific for TLS.

We forked the rustls library for cryptographic operations like deriving secrets. We also use it to encode and decode TLS messages.

The cryptographic library ring allows us to use the derived secrets to encrypt and decrypt TLS messages.

Modules

agent
Agents represent communication participants like Alice, Bob or Eve. Attackers are usually not represented by these Agents. Attackers are represented through a recipe term (see [InputAction]).

error

fuzzer
The fuzzer module setups the fuzzing loop. It also is responsible for gathering feedback from runs and restarting processes if they crash.

graphviz
This module adds plotting capabilities toTermsand Traces. The output of the functions in this module can be passed to the command line utility dot which is part of graphviz.

io
These are currently implemented by using an in-memory buffer. One might ask why we want two channels. There two very practical reasons for this. Note that these are advantages for the implementation and are not strictly required from a theoretical point of view.

term
The term module defines typedTermsof the form fn_add(x: u8, fn_square(y: u16)) → u16. Each function like fn_add or fn_square has a shape. The variables x and y each have a type. These types allow type checks during the runtime of the fuzzer. These checks restrict howTermscan be mutated in the fuzzer module.

tls
The tls module provides concrete implementations for the functions used in the term. The module offers a variety of [DynamicFunction]s which can be used in the fuzzing.

trace
This module contains Traces consisting of several Steps, of which each has either an OutputAction or InputAction. This is a declarative way of modeling communication between Agents. The TraceContext holds data, also known as VariableData, which is created by Agents during the concrete execution of the Trace. It also holds the Agents with the references to concrete PUT.

variable_data
Definition of the VariableData trait. A VariableData can contain any data which has a 'static type. This is true for [rustls::msgs::message::Message] for example.

Macros

define_signature

mutator

nyi_fn

term

term_arg