Module tlspuffin::agent

Agents represent communication participants like Alice, Bob or Eve. Attackers are usually not represented by these Agents. Attackers are represented through a recipe term (see [InputAction]).

Each Agent has an inbound and an outbound channel (see crate::io)

Structs

Agent
An Agent holds a non-cloneable reference to a Stream.

AgentDescriptor
AgentDescriptors act like a blueprint to spawn Agents with a corresponding server or client role and a specific TLs version. Essentially they are an Agent without a stream.

AgentName
Copyable reference to an Agent. It identifies exactly one agent.

Enums

TLSVersion