Module tlspuffin::algebra

The term module defines typedTermsof the form fn_add(x: u8, fn_square(y: u16)) → u16. Each function like fn_add or fn_square has a shape. The variables x and y each have a type. These types allow type checks during the runtime of the fuzzer. These checks restrict howTermscan be mutated in the fuzzer module.

Modules

atoms

This module provides an enum for terms. A term can either be a Variable or a Function. This also implements the serializability of terms.

dynamic_function

This module provides traits for calling rust functions dynamically. All functions which implement the DynamicFunction trait can be called by passing an array of Anys to it. The return value is again of type Any.

macros

This module provides a DLS for writing[Term]s within Rust.

signature

term 🔒

This module providesTermsas well as iterators over them.

Enums

Term

A first-order term: either a Variable or an application of an Function.

Traits

Subterms