Trait tlspuffin::concretize::PUT

pub trait PUT: Stream + Drop {
    fn new(c: Config) -> Result<Self, Error>
    where
        Self: Sized;

    fn progress(&mut self) -> Result<(), Error>;

    fn reset(&mut self) -> Result<(), Error>;

    fn register_claimer(
        &mut self, 
        claimer: Rc<RefCell<VecClaimer>>, 
        agent_name: AgentName
    );

    fn deregister_claimer(&mut self);

    fn change_agent_name(
        &mut self, 
        claimer: Rc<RefCell<VecClaimer>>, 
        agent_name: AgentName
    );

    fn describe_state(&self) -> &'static str;

    fn version(&self) -> &'static str;

    fn make_deterministic(&self);
}

Required methods

fn new(c: Config) -> Result<Self, Error> where
Self: Sized,

Create a new agent state for the PUT + set up buffers/BIOs

fn progress(&mut self) -> Result<(), Error>

Process incoming buffer, internal progress, can fill in output buffer

fn reset(&mut self) -> Result<(), Error>

In-place reset of the state

fn register_claimer(
&mut self,
claimer: Rc<RefCell<VecClaimer>>,
agent_name: AgentName
)

Register a new claim for agent_name

fn deregister_claimer(&mut self)

Remove all claims in self

fn change_agent_name(
&mut self,
claimer: Rc<RefCell<VecClaimer>>,
agent_name: AgentName
)

Change the agent name and the claimer of self

fn describe_state(&self) -> &'static str

Returns a textual representation of the state in which self is

fn version(&self) -> &'static str

Returns a textual representation of the version of the PUT used by self

fn make_deterministic(&self)

Make the PUT used by self determimistic in the future by making its PRNG “deterministic”