Module tlspuffin::fuzzer::libafl_setup

Statics

FRESH_ZOO_AFTER

MAX_ITERATIONS_PER_STAGE

Default value, how many iterations each stage gets, as an upper bound It may randomly continue earlier. Each iteration works on a different Input from the corpus

MAX_MUTATIONS_PER_ITERATION

MAX_TERM_SIZE

Above this term size we no longer mutate.

MAX_TRACE_LENGTH

MIN_TERM_SIZE

Below this term size we no longer mutate. Note that it is possible to reach smaller terms by having a mutation which removes all symbols in a single mutation.

MIN_TRACE_LENGTH

Functions

minimizer_feedback

no_feedback

no_minimizer_feedback

run_client

start

Starts the fuzzing loop

Type Definitions

ConcreteExecutor 🔒

ConcreteState 🔒