Module tlspuffin::fuzzer::seeds

Implementation of special traces. Each may represent a special TLS execution like a full handshake or an execution which crashes OpenSSL.

Functions

create_corpus

seed_client_attacker

seed_client_attacker12

seed_client_attacker_full
Seed which contains the whole transcript in the tree. This is rather huge >300 symbols

seed_cve_2021_3449

seed_freak

seed_heartbleed

seed_session_resumption_dhe

seed_session_resumption_dhe_full
Seed which contains the whole transcript in the tree. This is rather huge 10k symbols. It grows exponentially.

seed_session_resumption_ke

seed_successful

seed_successful12

seed_successful_mitm
Seed which triggers a MITM attack. It changes the cipher suite. This should fail.

seed_successful_with_ccs

seed_successful_with_tickets