Module tlspuffin::graphviz

This module adds plotting capabilities toTermsand Traces. The output of the functions in this module can be passed to the command line utility dot which is part of graphviz.

Constants

COLOR 🔒

COLOR_LEAVES 🔒

FONT 🔒

SHAPE 🔒

SHAPE_LEAVES 🔒

SHOW_LABELS 🔒

STYLE 🔒

Functions

write_graphviz