Trait tlspuffin::put::Put

pub trait Put: Stream + Drop + 'static {
    fn new(agent: &AgentDescriptor, config: PutConfig) -> Result<Self, Error>
    where
        Self: Sized;
    fn progress(&mut self, agent_name: &AgentName) -> Result<(), Error>;
    fn reset(&mut self, agent_name: AgentName) -> Result<(), Error>;
    fn config(&self) -> &PutConfig;
    fn rename_agent(&mut self, agent_name: AgentName) -> Result<(), Error>;
    fn describe_state(&self) -> &str;
    fn is_state_successful(&self) -> bool;
    fn version() -> &'static str
    where
        Self: Sized;
    fn make_deterministic()
    where
        Self: Sized;
    fn shutdown(&mut self) -> String;

    fn is_reusable_with(&self, other: &AgentDescriptor) -> bool { ... }
}

Required methods

fn new(agent: &AgentDescriptor, config: PutConfig) -> Result<Self, Error> where
Self: Sized,

Create a new agent state for the PUT + set up buffers/BIOs

fn progress(&mut self, agent_name: &AgentName) -> Result<(), Error>

Process incoming buffer, internal progress, can fill in output buffer

fn reset(&mut self, agent_name: AgentName) -> Result<(), Error>

In-place reset of the state

fn config(&self) -> &PutConfig

fn rename_agent(&mut self, agent_name: AgentName) -> Result<(), Error>

Propagate agent changes to the PUT

fn describe_state(&self) -> &str

Returns a textual representation of the state in which self is

fn is_state_successful(&self) -> bool

Checks whether the Put is in a good state

fn version() -> &'static str where
Self: Sized,

Returns a textual representation of the version of the PUT used by self

fn make_deterministic() where
Self: Sized,

Make the PUT used by self determimistic in the future by making its PRNG “deterministic”

fn shutdown(&mut self) -> String

Provided methods

fn is_reusable_with(&self, other: &AgentDescriptor) -> bool

checks whether a agent is reusable with the descriptor