Module tlspuffin::tls

The tls module provides concrete implementations for the functions used in the term. The module offers a variety of [DynamicFunction]s which can be used in the fuzzing.

Modules

error

fn_constants

fn_extensions
Extensions according to IANA: https://www.iana.org/assignments/tls-extensiontype-values/tls-extensiontype-values.xhtml

fn_fields

fn_impl
This modules contains all the concrete implementations of function symbols.

fn_messages
Extensions according to IANA: https://www.iana.org/assignments/tls-parameters/tls-parameters.xhtml#tls-parameters-7

fn_transcript

fn_utils

Statics

SIGNATURE
Signature which contains all functions defined in the tls module. A signature is responsible for linking function implementations to serialized data.