Module tlspuffin::tls::fn_impl::fn_utils

Functions

fn_append_certificate

fn_append_certificate_entry

fn_append_flight

fn_append_opaque_flight

fn_append_transcript

fn_certificate_entries_make

fn_decode_ecdh_pubkey

fn_decrypt_application

fn_decrypt_application_flight

Decrypt a whole flight of application messages and return a Vec of decrypted messages

fn_decrypt_handshake_flight

Decrypt a whole flight of handshake messages and return a Vec of decrypted messages

fn_decrypt_multiple_handshake_messages

Decrypt an Application data message containing multiple handshake messages and return a vec of handshake messages

fn_derive_binder

fn_derive_psk

fn_encode_ec_pubkey12

fn_encrypt12

fn_encrypt_application

fn_encrypt_handshake

fn_fill_binder

fn_find_encrypted_extensions

fn_find_server_certificate

fn_find_server_certificate_request

fn_find_server_certificate_verify

fn_find_server_finished

fn_find_server_ticket

fn_get_ticket

fn_get_ticket_age_add

fn_get_ticket_nonce

fn_named_group_secp256r1

fn_named_group_secp384r1

fn_named_group_x25519

fn_new_certificate

fn_new_certificate_entries

fn_new_certificates

fn_new_flight

fn_new_hrr_transcript

fn_new_opaque_flight

fn_new_pubkey12

fn_new_transcript

fn_new_transcript12

fn_no_psk

fn_psk

fn_u64_to_u32

suite_as_supported_suite