Enum tlspuffin::tls::rustls::error::Error

pub enum Error {

    InappropriateMessage {
        expect_types: Vec<ContentType>,
        got_type: ContentType,
    },
    InappropriateHandshakeMessage {
        expect_types: Vec<HandshakeType>,
        got_type: HandshakeType,
    },
    CorruptMessage,
    CorruptMessagePayload(CorruptMessagePayload),
    NoCertificatesPresented,
    UnsupportedNameType,
    DecryptError,
    EncryptError,
    PeerIncompatibleError(String),
    PeerMisbehavedError(String),
    AlertReceived(AlertDescription),
    InvalidCertificateEncoding,
    InvalidCertificateSignatureType,
    InvalidCertificateSignature,
    InvalidCertificateData(String),
    InvalidSct(Error),
    General(String),
    FailedToGetCurrentTime,
    FailedToGetRandomBytes,
    HandshakeNotComplete,
    PeerSentOversizedRecord,
    NoApplicationProtocol,
    BadMaxFragmentSize,

}

rustls reports protocol errors using this type.

Variants

InappropriateMessage

We received a TLS message that isn’t valid right now. expect_types lists the message types we can expect right now. got_type is the type we found. This error is typically caused by a buggy TLS stack (the peer or this one), a broken network, or an attack.

Fields

expect_types: Vec<ContentType>

Which types we expected

got_type: ContentType

What type we received

InappropriateHandshakeMessage

We received a TLS handshake message that isn’t valid right now. expect_types lists the handshake message types we can expect right now. got_type is the type we found.

Fields

expect_types: Vec<HandshakeType>

Which handshake type we expected

got_type: HandshakeType

What handshake type we received

CorruptMessage

The peer sent us a syntactically incorrect TLS message.

CorruptMessagePayload(CorruptMessagePayload)

The peer sent us a TLS message with invalid contents.

NoCertificatesPresented

The peer didn’t give us any certificates.

UnsupportedNameType

The certificate verifier doesn’t support the given type of name.

DecryptError

We couldn’t decrypt a message. This is invariably fatal.

EncryptError

We couldn’t encrypt a message because it was larger than the allowed message size. This should never happen if the application is using valid record sizes.

PeerIncompatibleError(String)

The peer doesn’t support a protocol version/feature we require. The parameter gives a hint as to what version/feature it is.

PeerMisbehavedError(String)

The peer deviated from the standard TLS protocol. The parameter gives a hint where.

AlertReceived(AlertDescription)

We received a fatal alert. This means the peer is unhappy.

InvalidCertificateEncoding

We received an invalidly encoded certificate from the peer.

InvalidCertificateSignatureType

We received a certificate with invalid signature type.

InvalidCertificateSignature

We received a certificate with invalid signature.

InvalidCertificateData(String)

We received a certificate which includes invalid data.

InvalidSct(Error)

The presented SCT(s) were invalid.

General(String)

A catch-all error for unlikely errors.

FailedToGetCurrentTime

We failed to figure out what time it currently is.

FailedToGetRandomBytes

We failed to acquire random bytes from the system.

HandshakeNotComplete

This function doesn’t work until the TLS handshake is complete.

PeerSentOversizedRecord

The peer sent an oversized record/fragment.

NoApplicationProtocol

An incoming connection did not support any known application protocol.

BadMaxFragmentSize

The max_fragment_size value supplied in configuration was too small, or too large.

Implementations

impl Error

pub fn corrupt_message(kind: ContentType) -> Self