Module tlspuffin::tls::rustls::kx

Structs

KeyExchange

An in-progress key exchange. This has the algorithm, our private key, and our public key.

SupportedKxGroup

A key-exchange group supported by rustls.

Statics

ALL_KX_GROUPS

A list of all the key exchange groups supported by rustls.

SECP256R1

Ephemeral ECDH on secp256r1 (aka NIST-P256)

SECP384R1

Ephemeral ECDH on secp384r1 (aka NIST-P384)

X25519

Ephemeral ECDH on curve25519 (see RFC7748)