Enum tlspuffin::tls::rustls::msgs::enums::SignatureAlgorithm

pub enum SignatureAlgorithm {
    Anonymous,
    RSA,
    DSA,
    ECDSA,
    ED25519,
    ED448,
    Unknown(u8),
}

The SignatureAlgorithm TLS protocol enum.

Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

Variants

Anonymous

RSA

DSA

ECDSA

ED25519

ED448

Unknown(u8)

Implementations

impl SignatureAlgorithm

pub fn get_u8(&self) -> u8