Struct tlspuffin::tls::rustls::msgs::message::OpaqueMessage

pub struct OpaqueMessage {
    pub typ: ContentType,
    pub version: ProtocolVersion,
    pub payload: Payload,
}

A TLS frame, named TLSPlaintext in the standard.

This type owns all memory for its interior parts. It is used to read/write from/to I/O buffers as well as for fragmenting, joining and encryption/decryption. It can be converted into a Message by decoding the payload.

Fields
typ: ContentTypeversion: ProtocolVersionpayload: Payload
Implementations

impl OpaqueMessage

pub fn read(r: &mut Reader<'_>) -> Result<Self, MessageError>

MessageError allows callers to distinguish between valid prefixes (might become valid if we read more data) and invalid data.

pub fn encode(self) -> Vec<u8>

pub fn into_plain_message(self) -> PlainMessage

Force conversion into a plaintext message.

This should only be used for messages that are known to be in plaintext. Otherwise, the OpaqueMessage should be decrypted into a PlainMessage using a MessageDecrypter.

const MAX_PAYLOAD: u16 = 18_432u16

This is the maximum on-the-wire size of a TLSCiphertext. That’s 2^14 payload bytes, a header, and a 2KB allowance for ciphertext overheads.

const HEADER_SIZE: u16 = 5u16

Content type, version and size.

pub const MAX_WIRE_SIZE: usize = 18_437usize

Maximum on-wire message size.