Enum tlspuffin::tls::rustls::quic::KeyChange

pub enum KeyChange {
    Handshake {
        keys: Keys,
    },
    OneRtt {
        keys: Keys,
        next: Secrets,
    },
}

Key material for use in QUIC packet spaces

QUIC uses 4 different sets of keys (and progressive key updates for long-running connections):

Initial: these can be created from Keys::initial()

0-RTT keys: can be retrieved from QuicExt::zero_rtt_keys()

Handshake: these are returned from QuicExt::write_hs() after ClientHello and ServerHello messages have been exchanged

1-RTT keys: these are returned from QuicExt::write_hs() after the handshake is done

Once the 1-RTT keys have been exchanged, either side may initiate a key update. Progressive update keys can be obtained from the Secrets returned in KeyChange::OneRtt. Note that only packet keys are updated by key updates; header protection keys remain the same.

Variants

Handshake

Keys for the handshake space

Fields

keys: Keys

Header and packet keys for the handshake space

OneRtt

Keys for 1-RTT data

Fields

keys: Keys

Header and packet keys for 1-RTT data

next: Secrets

Secrets to derive updated keys from