Struct tlspuffin::tls::rustls::tls13::key_schedule::KeySchedule

pub struct KeySchedule {
    current: Prk,
    algorithm: Algorithm,
}

This is the TLS1.3 key schedule. It stores the current secret and the type of hash. This isn’t used directly; but only through the typestates.

Fields
current: Prkalgorithm: Algorithm
Implementations

impl KeySchedule

fn new(algorithm: Algorithm, secret: &[u8]) -> Self

fn algorithm(&self) -> Algorithm

fn new_with_empty_secret(algorithm: Algorithm) -> Self

fn input_empty(&mut self)

Input the empty secret.

fn input_secret(&mut self, secret: &[u8])

Input the given secret.

fn derive<T, L>(&self, key_type: L, kind: SecretKind, hs_hash: &[u8]) -> T where
T: for<'a> From<Okm<'a, L>>,
L: KeyType,

Derive a secret of given kind, using current handshake hash hs_hash.

fn derive_logged_secret(
&self,
kind: SecretKind,
hs_hash: &[u8],
key_log: &dyn KeyLog,
client_random: &[u8; 32]
) -> Prk

fn derive_for_empty_hash<T>(&self, kind: SecretKind) -> T where
T: for<'a> From<Okm<'a, Algorithm>>,

Derive a secret of given kind using the hash of the empty string for the handshake hash. Useful only for SecretKind::ResumptionPSKBinderKey and SecretKind::DerivedSecret.

fn sign_finish(&self, base_key: &Prk, hs_hash: &Digest) -> Tag

Sign the finished message consisting of hs_hash using a current traffic secret.

fn sign_finish_raw(&self, base_key: &Prk, hs_hash: &[u8]) -> Tag

fn sign_verify_data(&self, base_key: &Prk, hs_hash: &Digest) -> Tag

Sign the finished message consisting of hs_hash using the key material base_key.

fn sign_verify_data_raw(&self, base_key: &Prk, hs_hash: &[u8]) -> Tag

fn derive_next(&self, base_key: &Prk) -> Prk

Derive the next application traffic secret, returning it.

fn derive_ticket_psk(&self, rms: &Prk, nonce: &[u8]) -> Vec<u8>

Derive the PSK to use given a resumption_master_secret and ticket_nonce.

fn export_keying_material(
&self,
current_exporter_secret: &Prk,
out: &mut [u8],
label: &[u8],
context: Option<&[u8]>
) -> Result<(), Error>