Struct tlspuffin::tls::rustls::vecbuf::ChunkVecBuffer

pub struct ChunkVecBuffer {
    chunks: VecDeque<Vec<u8>>,
    limit: Option<usize>,
}

This is a byte buffer that is built from a vector of byte vectors. This avoids extra copies when appending a new byte vector, at the expense of more complexity when reading out.

Fields
chunks: VecDeque<Vec<u8>>limit: Option<usize>
Implementations

impl ChunkVecBuffer

pub fn new(limit: Option<usize>) -> Self

pub fn set_limit(&mut self, new_limit: Option<usize>)

Sets the upper limit on how many bytes this object can store.

Setting a lower limit than the currently stored data is not an error.

A None limit is interpreted as no limit.

pub fn is_empty(&self) -> bool

If we’re empty

pub fn len(&self) -> usize

How many bytes we’re storing

pub fn apply_limit(&self, len: usize) -> usize

For a proposed append of len bytes, how many bytes should we actually append to adhere to the currently set limit?

pub fn append_limited_copy(&mut self, bytes: &[u8]) -> usize

Append a copy of bytes, perhaps a prefix if we’re near the limit.

pub fn append(&mut self, bytes: Vec<u8>) -> usize

Take and append the given bytes.

pub fn pop(&mut self) -> Option<Vec<u8>>

Take one of the chunks from this object. This function panics if the object is_empty.

pub fn read(&mut self, buf: &mut [u8]) -> Result<usize>

Read data out of this object, writing it into buf and returning how many bytes were written there.

fn consume(&mut self, used: usize)

pub fn write_to(&mut self, wr: &mut dyn Write) -> Result<usize>

Read data out of this object, passing it wr