Struct tlspuffin::tls::rustls::verify::AllowAnyAnonymousOrAuthenticatedClient

pub struct AllowAnyAnonymousOrAuthenticatedClient {
    inner: AllowAnyAuthenticatedClient,
}

A ClientCertVerifier that will allow both anonymous and authenticated clients, without any name checking.

Client authentication will be requested during the TLS handshake. If the client offers a certificate then this acts like AllowAnyAuthenticatedClient, otherwise this acts like NoClientAuth.

Fields
inner: AllowAnyAuthenticatedClient
Implementations

impl AllowAnyAnonymousOrAuthenticatedClient

pub fn new(roots: RootCertStore) -> Arc<dyn ClientCertVerifier>

Construct a new AllowAnyAnonymousOrAuthenticatedClient.

roots is the list of trust anchors to use for certificate validation.