Struct tlspuffin::tls::rustls::verify::AllowAnyAuthenticatedClient

pub struct AllowAnyAuthenticatedClient {
    roots: RootCertStore,
}

A ClientCertVerifier that will ensure that every client provides a trusted certificate, without any name checking.

Fields
roots: RootCertStore
Implementations

impl AllowAnyAuthenticatedClient

pub fn new(roots: RootCertStore) -> Arc<dyn ClientCertVerifier>

Construct a new AllowAnyAuthenticatedClient.

roots is the list of trust anchors to use for certificate validation.