Struct tlspuffin::tls::rustls::verify::WebPkiVerifier

pub struct WebPkiVerifier {
    roots: RootCertStore,
    ct_policy: Option<CertificateTransparencyPolicy>,
}

Default ServerCertVerifier, see the trait impl for more information.

Fields
roots: RootCertStorect_policy: Option<CertificateTransparencyPolicy>
Implementations

impl WebPkiVerifier

pub fn new( roots: RootCertStore, ct_policy: Option<CertificateTransparencyPolicy>, ) -> Self

Constructs a new WebPkiVerifier.

roots is the set of trust anchors to trust for issuing server certs.

ct_logs is the list of logs that are trusted for Certificate Transparency. Currently CT log enforcement is opportunistic; see https://github.com/rustls/rustls/issues/479.

pub fn verification_schemes() -> Vec<SignatureScheme>

Returns the signature verification methods supported by webpki.