Module tlspuffin::tls::seeds

Implementation of special traces. Each may represent a special TLS execution like a full handshake or an execution which crashes OpenSSL.

Macros

corpus 🔒

Traits

SeedExecutor

SeedHelper

Functions

_seed_client_attacker12 🔒

_seed_client_attacker_full 🔒

Seed which contains the whole transcript in the tree. This is rather huge >300 symbols

create_corpus

seed_client_attacker

seed_client_attacker12

seed_client_attacker_auth

seed_client_attacker_full

seed_cve_2021_3449

seed_cve_2022_25638

https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-25638

seed_cve_2022_25640

https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-25640

seed_cve_2022_25640_simple

A simplified version of seed_cve_2022_25640

seed_freak

seed_heartbleed

seed_server_attacker_full

seed_session_resumption_dhe

seed_session_resumption_dhe_full

Seed which contains the whole transcript in the tree. This is rather huge 10k symbols. It grows exponentially.

seed_session_resumption_ke

seed_successful

seed_successful12

seed_successful12_with_tickets

seed_successful_client_auth

seed_successful_mitm

Seed which triggers a MITM attack. It changes the cipher suite. This should fail.

seed_successful_with_ccs

seed_successful_with_tickets