Struct tlspuffin::trace::TraceContext

pub struct TraceContext {
    pub claimer: Rc<RefCell<VecClaimer>>,
    // some fields omitted
}

The TraceContext contains a list of VariableData, which is known as the knowledge of the attacker. VariableData can contain data of various types like for example client and server extensions, cipher suits or session ID It also holds the concrete references to the Agents and the underlying streams, which contain the messages which have need exchanged and are not yet processed by an output step.

Fields
claimer: Rc<RefCell<VecClaimer>>
Implementations

impl TraceContext

pub fn new() -> Self

pub fn add_knowledge(&mut self, knowledge: Knowledge)

pub fn number_matching_message(
&self,
agent: AgentName,
type_id: TypeId,
tls_message_type: Option<TlsMessageType>
) -> u16

Count the number of sub-messages of type [type_id] in the output message [in_step_id].

pub fn find_variable(
&self,
query_type_shape: TypeShape,
query: Query
) -> Option<&dyn VariableData>

Returns the variable which matches best -> highest specificity If we want a variable with lower specificity, then we can just query less specific

pub fn add_to_inbound(
&mut self,
agent_name: AgentName,
message: &OpaqueMessage
) -> Result<(), Error>

Adds data to the inbound Channel of the Agent referenced by the parameter “agent”.

pub fn next_state(&mut self, agent_name: AgentName) -> Result<(), Error>

pub fn take_message_from_outbound(
&mut self,
agent_name: AgentName
) -> Result<Option<MessageResult>, Error>

Takes data from the outbound Channel of the Agent referenced by the parameter “agent”. See [MemoryStream::take_message_from_outbound]

pub fn new_openssl_agent(
&mut self,
descriptor: &AgentDescriptor
) -> Result<AgentName, Error>

pub fn find_agent(&self, name: AgentName) -> Result<&Agent, Error>

pub fn reset_agents(&mut self)