Module tlspuffin::tls

The tls module provides concrete implementations for the functions used in the term. The module offers a variety of [DynamicFunction]s which can be used in the fuzzing.

Modules

fn_impl

This modules contains all the concrete implementations of function symbols.

key_exchange 🔒

key_schedule 🔒

rustls

Extracted from rustls fork. Upstreaming this fork is becoming unfeasible and there are not benefits in keeping up with the latest version of rustls. If we want to support fuzzing new features of upcomping TLS versions then we have to manually integrate them.

seeds

Implementation of special traces. Each may represent a special TLS execution like a full handshake or an execution which crashes OpenSSL.

violation

Statics

TLS_SIGNATURE

Signature which contains all functions defined in the tls module. A signature is responsible for linking function implementations to serialized data.

Functions

fn_debug 🔒

Function symbol which can be used for debugging