Module tlspuffin::tls::rustls

Extracted from rustls fork. Upstreaming this fork is becoming unfeasible and there are not benefits in keeping up with the latest version of rustls. If we want to support fuzzing new features of upcomping TLS versions then we have to manually integrate them.

This module contains primitives required to perform TLS 1.2 and 1.3 handshakes.

License of rustls:

Copyright (c) 2016 Joseph Birr-Pixton <jpixton@gmail.com>

Permission is hereby granted, free of charge, to any
person obtaining a copy of this software and associated
documentation files (the "Software"), to deal in the
Software without restriction, including without
limitation the rights to use, copy, modify, merge,
publish, distribute, sublicense, and/or sell copies of
the Software, and to permit persons to whom the Software
is furnished to do so, subject to the following
conditions:

The above copyright notice and this permission notice
shall be included in all copies or substantial portions
of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF
ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
DEALINGS IN THE SOFTWARE.

Modules

anchors

cipher

client

Items for use in a client.

conn

error

hash_hs

key

key_log

kx

limited_cache

msgs

quic

APIs for implementing QUIC TLS

rand

record_layer

server

Items for use in a server.

sign

Message signing interfaces and implementations.

suites

ticketer

tls12

tls13

vecbuf

verify

versions

x509