Module tlspuffin::tls::rustls::quic

APIs for implementing QUIC TLS

Structs

DirectionalKeys

Keys used to communicate in a single direction

HeaderProtectionKey

A QUIC header protection key

Keys

Complete set of keys used to communicate with the peer

PacketKey

Keys to encrypt or decrypt the payload of a packet

PacketKeySet

Packet protection keys for bidirectional 1-RTT communication

Secrets

Secrets used to encrypt/decrypt traffic

Tag

AEAD tag, must be appended to encrypted cipher text

Enums

KeyChange

Key material for use in QUIC packet spaces

Version

QUIC protocol version

Traits

QuicExt

Generic methods for QUIC sessions

Functions

nonce_for 🔒

Compute the nonce to use for encrypting or decrypting packet_number