Module tlspuffin::tls::rustls::tls13::key_schedule

Structs

KeySchedule

This is the TLS1.3 key schedule. It stores the current secret and the type of hash. This isn’t used directly; but only through the typestates.

KeyScheduleEarly

KeySchedule for early data stage.

KeyScheduleHandshake

KeyScheduleHandshakeStart

KeySchedule during handshake.

KeySchedulePreHandshake

Pre-handshake key schedule

KeyScheduleTraffic

KeySchedule during traffic stage. All traffic & exporter keys are guaranteed to be available.

KeyScheduleTrafficWithClientFinishedPending

KeySchedule during traffic stage, retaining the ability to calculate the client’s finished verify_data. The traffic stage key schedule can be extracted from it through signing the client finished hash.

PayloadU8Len

Enums

SecretKind 🔒

The kinds of secret we can extract from KeySchedule.

Functions

derive_traffic_iv

derive_traffic_key

hkdf_expand

hkdf_expand_info 🔒