Module tlspuffin::tls::rustls::verify

Structs

AllowAnyAnonymousOrAuthenticatedClient

A ClientCertVerifier that will allow both anonymous and authenticated clients, without any name checking.

AllowAnyAuthenticatedClient

A ClientCertVerifier that will ensure that every client provides a trusted certificate, without any name checking.

CertificateTransparencyPolicy

Policy for enforcing Certificate Transparency.

ClientCertVerified

Zero-sized marker type representing verification of a client cert chain.

DnsName

A type which encapsuates a string that is a syntactically valid DNS name.

FinishedMessageVerified

HandshakeSignatureValid

Zero-sized marker type representing verification of a signature.

NoClientAuth

Turns off client authentication.

ServerCertVerified

Zero-sized marker type representing verification of a server cert chain.

WebPkiVerifier

Default ServerCertVerifier, see the trait impl for more information.

Statics

ECDSA_SHA256 🔒

ECDSA_SHA384 🔒

ED25519 🔒

RSA_PSS_SHA256 🔒

RSA_PSS_SHA384 🔒

RSA_PSS_SHA512 🔒

RSA_SHA256 🔒

RSA_SHA384 🔒

RSA_SHA512 🔒

SUPPORTED_SIG_ALGS 🔒

Which signature verification mechanisms we support. No particular order.

Traits

ClientCertVerifier

Something that can verify a client certificate chain

ServerCertVerifier

Something that can verify a server certificate chain, and verify signatures made by certificates.

Functions

construct_tls13_client_verify_message

Constructs the signature message specified in section 4.4.3 of RFC8446.

construct_tls13_client_verify_message_raw

Constructs the signature message specified in section 4.4.3 of RFC8446.

construct_tls13_server_verify_message

Constructs the signature message specified in section 4.4.3 of RFC8446.

construct_tls13_server_verify_message_raw

Constructs the signature message specified in section 4.4.3 of RFC8446.

construct_tls13_verify_message 🔒

convert_alg_tls13 🔒

convert_scheme 🔒

pki_error 🔒

prepare 🔒

unix_time_millis 🔒

verify_sig_using_any_alg 🔒

verify_signed_struct 🔒

verify_tls13 🔒

Type Aliases

CertChainAndRoots 🔒

SignatureAlgorithms 🔒